Add files via upload
Maintain two sequences of distinct integers, one in ascending order (denoted as linked list A) and one in descending order (denoted as linked list B). Given a positive integer k, we may pick one integer a from A, and one integer b from B, such that a + b = k. The goal is to find out how many such pairs exist for the given k and linked lists A and B.